code refactor and data file rename
